requirements.txt
asgiref==3.7.2
distlib==0.3.8
Django==5.0
djangorestframework==3.15.1
filelock==3.14.0
pillow==10.3.0
platformdirs==4.2.2
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.26.2
